% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_clock.R, R/clock_chart.R
\name{basic_clock}
\alias{basic_clock}
\title{Give a data set containing times in 24 hours format (19:30, for example)}
\usage{
basic_clock()

basic_clock()
}
\arguments{
\item{data}{A data set}

\item{time}{Time in 24 hours format in the data set}
}
\value{
A ggplot data clock

A ggplot data clock
}
\description{
This function will plot those times on a 24 hour clock to show which
events took place at what times. The line segments are drawn using
trigonometric angles and complex equations.

This function will plot those times on a 24 hour clock to show which
events took place at what times. The line segments are drawn using
trigonometric angles and complex equations.
}
\examples{
basic_clock()
clock_chart(data = df, time = time)
}
