% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdquake}
\alias{bdquake}
\title{Data of Earthquakes in and around Bangladesh}
\format{
A data frame with 13 rows and 5 variables:
\describe{
\item{latitude}{Latitude}
\item{longitude}{Longitude}
\item{depth}{Depth}
\item{mag}{Magnitude}
\item{hms}{Time}
}
}
\source{
\href{https://earthquake.usgs.gov/earthquakes/search/}{USGS}
}
\usage{
bdquake
}
\description{
A dataset containing earthquakes magnitude, dpeth, and location information.
}
\keyword{datasets}
