% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chatdf}
\alias{chatdf}
\title{Time of chat by two individuals}
\format{
A data frame with 25 rows and 2 variables:
\describe{
\item{name}{Name of the chat participant}
\item{time}{Time in HH:MM:SS format}
\item{turn}{Turn of the chat}
...
}
}
\source{
\url{#}
}
\usage{
chatdf
}
\description{
A dataset containing timings of chat by two individuals Abid and Abir
The set contains 25 pushes by each individual
}
\keyword{datasets}
