% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_chart_len.R
\name{clock_chart_len}
\alias{clock_chart_len}
\title{Clock Chart with Modified Length}
\usage{
clock_chart_len(data, time, crit, Col = "black")
}
\arguments{
\item{data}{A data frame}

\item{time}{Time in 24 hours. The allowed time formats for these family
of charts are \code{HH:MM:SS}, \code{HH:MM} or even \code{H:M} (such as \verb{12;30:09}
or \code{9:3}).}

\item{crit}{A numeric vector by which length of hands will be modified.}

\item{Col}{An optional color name for the hands, defaulted to \code{black}.}
}
\description{
This function will plot time of events on a 24 hour clock to show which
events took place at what times. The length of the hands are modified
according to a numeric vector. A color name can be used for the hands.
This function has been seen to be unnecessary and its purpose is better-served by \code{clock_chart_qnt()}, and hence will be deprecated in the next version..
}
\details{
Change the title, subtitle or the caption of the plot with
\code{ggplot2::labs()}. See examples.
}
\examples{
p1 <- clock_chart_len(data = bdquake, time = hms, crit = depth)
p1 + ggplot2::labs(title = "Earthquakes in Bangladesh since 2023")
}
\seealso{
\code{\link[=clock_chart_col]{clock_chart_col()}} for coloring by a numeric variable,
\code{\link[=clock_chart_qnt]{clock_chart_qnt()}} for coloring and modifying length by a numeric variable,
\code{\link[=clock_chart_qlt]{clock_chart_qlt()}} for coloring by a qualitative variable,
\code{\link[=clock_chart]{clock_chart()}} for the simplest clock chart
}
