% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_chart_qlt.R
\name{clock_chart_qlt}
\alias{clock_chart_qlt}
\title{Clock Chart, Hands Colored by a Factor}
\usage{
clock_chart_qlt(data, time, crit)
}
\arguments{
\item{data}{A data frame}

\item{time}{Time in 24 hours. The allowed time formats for these family
of charts are \code{HH:MM:SS}, \code{HH:MM} or even \code{H:M} (such as \verb{12;30:09}
or \code{9:3}).}

\item{crit}{The qualitative vector by which hands will be colored.}
}
\description{
This function will plot time of events on a 24 hour clock to show which
events took place at what times. The hands are colored by a
qualitative (factor) vector.
}
\details{
Change the title, subtitle or the caption of the plot with
\code{ggplot2::labs()}. Change the legend title by adding
\code{ggplot2::labs(color = "TITLE")}. Add
or modify legend by \code{theme(legend.position = "POSITION")}; the valid
position names in \code{ggplot2} are \code{top}, \code{bottom}, \code{right}, and \code{left},
excluding more complex options.
}
\examples{
# A plot showing sms receiving times based on
# criteria (type/sender/invoked)
clock_chart_qlt(smsclock, time = time, crit = sender)+
ggplot2::labs(color = "Sender", title = "SMS's Received throughout th Day")
}
\seealso{
\code{\link[=clock_chart_col]{clock_chart_col()}} for coloring by a numeric variable,
\code{\link[=clock_chart_qnt]{clock_chart_qnt()}} for coloring and modifying length by a numeric variable,
\code{\link[=clock_chart]{clock_chart()}} for the simplest clock chart
}
