% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clockPlot.R
\name{clockPlot}
\alias{clockPlot}
\title{clockPlot}
\usage{
clockPlot(
  time,
  column = "",
  secondHand = FALSE,
  labels = "numeric",
  minuteMarks = TRUE,
  hourMarks = TRUE,
  border = TRUE
)
}
\arguments{
\item{time}{Either a tibble with a column of type POSIX or a single value of type POSIX (time is what we are plotting)}

\item{column}{The name of the column if time is a tibble as type String (because it is defining which column is needed)}

\item{secondHand}{Should the secondhand be shown. Default of FALSE means it is not shown (it controls the second hand visibility)}

\item{labels}{What labels to show for each hour, if any as type String (it controls the hour number style)}

\item{minuteMarks}{Should the marks be shown for each minute. Default of TRUE means they are shown (it controls the minute marks visibility)}

\item{hourMarks}{Should the marks be shown for each hour. Default of TRUE means they are shown (it controls the hour marks visibility)}

\item{border}{Should there be a border around the clock. Default of TRUE means there is (it controls the border visibility)}
}
\value{
A single clockPlot or the tibble with a clockPlot column in the first position
}
\description{
Uses ggplot to create a clock face graphic for visual purposes. This could then be saved and used as a visual aid in a document or in print
}
\examples{
clockPlot(Sys.time())
clockPlot(Sys.time(), labels = "roman")

}
