% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{conv_hms}
\alias{conv_hms}
\title{Convert \code{HH:MM} to \code{HH:MM:SS}}
\usage{
conv_hms(df, time, new)
}
\arguments{
\item{df}{A data set}

\item{time}{a \code{time} or \code{character} variable in the format \code{HH:MM}}

\item{new}{New name of the variable}
}
\value{
Time in HH:MM format
}
\description{
This function converts \code{HH:MM} to \code{HH:MM:SS}. \code{19:30}, for example, is
turned into \code{19:30:00}. This prepares the times for plotting, in case you
do not have it in \code{HH:MM:SS} format.
}
\details{
The converted \code{HH:MM:SS} actually gives \code{HH:MM:00} format, the second part
made \code{00}, since in practice it has negligible effect on the plot.
}
\examples{
df <- data.frame(time = c("06:00", "08:00", "17:30"))
df = conv_hms(df, time, New)
df
}
