% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_clock.R
\name{data_clock}
\alias{data_clock}
\title{Give a data set containing times in 24 hours format}
\usage{
data_clock(Data, Time, Criteria)
}
\arguments{
\item{Data}{A data set}

\item{Time}{Time in 24 hours format in the data set}

\item{Criteria}{The variable by which to change color/width of time lines}
}
\value{
A ggplot data clock
}
\description{
This function will find plot those times on a 24 hour clock to show which
events took place at what times. The line segments are drawn using
trigonometric angles and complex equations.
}
