% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{dataclock-package}
\alias{dataclock}
\alias{dataclock-package}
\title{dataclock: Show event time on a 24 hours clock}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
