% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gitcommit}
\alias{gitcommit}
\title{Data of git commits in this repository}
\format{
A data frame with 110 rows and 6 variables:
\describe{
\item{id}{Commit ID}
\item{weekday}{Week Day}
\item{month}{Month}
\item{day}{Day of the month}
\item{time}{Time--this is our desired column}
\item{year}{Year}
}
}
\source{
\href{clockplot repository}{Weather Base}
}
\usage{
gitcommit
}
\description{
A dataset containing commit information of all the commits in this
repository (\code{clockplot}) up to 12 Sep, 2024. Some columns, including the
commit message have been omitted.
}
\keyword{datasets}
