% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{rm_sec}
\alias{rm_sec}
\title{Remove second part from time value
This function removes second part from time value}
\usage{
rm_sec(df, time)
}
\arguments{
\item{time}{a variable in the format `HH:MM:SS`}

\item{data}{A data set}
}
\value{
Time in HH:MM format
}
\description{
Remove second part from time value
This function removes second part from time value
}
\examples{
df <- data.frame(time = c("06:00:00", "08:00:00", "17:30:00"))
rm_sec(df, time)
}
