% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_clock.R
\name{basic_clock}
\alias{basic_clock}
\title{Give a data set containing times in 24 hours format (19:30, for example)}
\usage{
basic_clock()
}
\value{
A ggplot clock chart
}
\description{
This function will plot those times on a 24 hour clock to show which
events took place at what times. The line segments are drawn using
trigonometric angles and complex equations.
}
