% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bdtemp}
\alias{bdtemp}
\title{Data of Average Monthly Temperature of Bangladesh Divisional Cities}
\format{
A data frame with 36 rows and 3 variables:
\describe{
\item{Temperature}{Average monthly temperature}
\item{Month}{Month of the year}
\item{City}{Name of the city}
...
}
}
\source{
\href{https://www.weatherbase.com/weather/weather.php3?s=032914&refer=/}{Weather Base}
}
\usage{
bdtemp
}
\description{
A dataset containing temperature of Dhaka, Sylhet, and Chittagong Cities.
}
\keyword{datasets}
