% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brintcity}
\alias{brintcity}
\title{Data of Bangladesh Railway Express Train Schedule}
\format{
A data frame with 84 rows and 10 variables:
\describe{
\item{Sl}{Serial}
\item{Train_no}{Train No.}
\item{Name}{Train name}
\item{Type}{Type of train}
\item{Off_day}{The day of the week when the train does not run}
\item{Origin}{The station from where the train departs}
\item{Departure}{The time of departure}
\item{Destination}{The station where the train is heading}
\item{Arrival}{The time of arrival}
\item{Zone}{East/West}
}
}
\source{
\href{https://railway.portal.gov.bd/sites/default/files/files/railway.portal.gov.bd/page/e64d9448_0615_4316_87f0_deb10f5c847d/Intercity\%20Trains\%20.pdf}{Bangladesh Railway}
}
\usage{
brintcity
}
\description{
A dataset containing Bangladesh Railway Express Train schedule
}
\keyword{datasets}
