% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_chart.R
\name{clock_chart}
\alias{clock_chart}
\title{Give a data set containing times in 24 hours format}
\usage{
clock_chart(data, time)
}
\arguments{
\item{data}{A data frame}

\item{time}{Time in 24 hours HH:MM format in the data set  (19:30, for example)}
}
\value{
A ggplot data clock
}
\description{
This function will plot those times on a 24 hour clock to show which
events took place at what times. The line segments are drawn using
trigonometric angles and complex equations.
}
\examples{
df <- data.frame(time = c("06:00:00", "08:00:00", "17:30:00"))
clock_chart(df, time)
}
