% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_chart_col.R
\name{clock_chart_col}
\alias{clock_chart_col}
\title{Give a data set containing times in 24 hours format (HH:MM:SS)}
\usage{
clock_chart_col(data, time, crit, high = "red", low = "green")
}
\arguments{
\item{data}{A data frame}

\item{time}{Time in 24 hours HH:MM format in the data set  (19:30, for example)}

\item{crit}{a quantitative vector by which lines will be colored.}

\item{high}{color of high values. The default is `red`}

\item{low}{color of high values. The default is `green`.
The color names can be vice versa or other colors, depending on the context.}
}
\value{
A ggplot data clock with colored hands.
}
\description{
This function will plot those times on a 24 hour clock to show which
events took place at what times. The line segments are drawn using
trigonometric angles and complex equations. The lines are colored based on a variable.
}
\examples{
df <- data.frame(time = c("06:00:00", "08:00:00", "17:30:00"),
crit = c(3,6,9))
clock_chart_col(df, time, crit)
}
