% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_chart_col.R
\name{clock_chart_col}
\alias{clock_chart_col}
\title{A Clock Chart, hands colored by a quantitative variable}
\usage{
clock_chart_col(data, time, crit, high = "red", low = "green")
}
\arguments{
\item{data}{A data frame}

\item{time}{Time in 24 hours HH:MM:SS format in the data set  (19:30:01, for example)}

\item{crit}{a quantitative vector by which lines will be colored.}

\item{high}{The color name for the high values. The default is `red`}

\item{low}{The color name for the high values. The default is `green`.
The color names can be vice versa or other colors, depending on the context.}
}
\value{
A ggplot data clock with colored hands.
}
\description{
This function will plot time of events on a 24 hour clock to show which
events took place at what times. The lines are colored by a criteria.
}
\details{
To
}
\examples{
df <- data.frame(time = c("06:00:00", "08:00:00", "17:30:00"),
crit = c(3,6,9))
clock_chart_col(df, time, crit = crit, high = "blue", low = "red")
}
