% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_chart_len.R
\name{clock_chart_len}
\alias{clock_chart_len}
\title{Clock Chart with Modified length}
\usage{
clock_chart_len(data, time, len, col = "black")
}
\arguments{
\item{data}{A data frame}

\item{time}{Time in 24 hours HH:MM:SS format in the data set  (19:30:01, for example)}

\item{len}{A quantitative vector by which length of hands will be modified.}

\item{col}{An optional color name for the hands, defaulted to `black`.}
}
\value{
A ggplot data clock with modified hands.
}
\description{
This function will plot time of events on a 24 hour clock to show which
events took place at what times. The length of the hands are modified
according to a quantitative vector. A color name can be used for the hands.
For better use of color and length together, consider using `clock_chart_qnt`.
}
\examples{
df <- data.frame(time = c("06:00:00", "08:00:00", "17:30:00"),
value = c(3,6,9))
clock_chart_len(data = df, time = time, len = value, col = "blue")
}
