% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_chart_qlt.R
\name{clock_chart_qlt}
\alias{clock_chart_qlt}
\title{Clock Chart with Colored Hands}
\usage{
clock_chart_qlt(data, time, crit)
}
\arguments{
\item{data}{A data frame}

\item{time}{Time in 24 hours HH:MM:SS format in the data set  (19:30:01, for example)}

\item{crit}{The qualitative vector by which hands will be colored.}
}
\value{
A ggplot clock chart with hands colored by a criteria.
}
\description{
This function will plot time of events on a 24 hour clock to show which
events took place at what times. The hands are colored by a
qualitative vector.
}
\examples{
# A plot showing sms receiving times based on criteria (type/sender/invoked)
clock_chart_qlt(smsclock, time = time, crit = sender)
df <- tibble::tibble(hr = sample(0:23, 50, replace = TRUE),
                  mnt = sample(0:59, 50, replace = TRUE),
                  sec = sample(0:59, 50, replace = TRUE),
                  time = paste(hr, mnt, sec, sep = ":"),
                  value = factor(sample(5,50, replace = TRUE)))
clock_chart_qlt(df, time, crit = value)
p1 <- clock_chart_qlt(df, time, crit = value)
p1 + ggplot2::theme(legend.position = "right")+
ggplot2::labs(title = "Clock chart of factor values")
data(chatdf)
clock_chart_qlt(chatdf, time = time, crit = name)
}
