% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_chart_qnt.R
\name{clock_chart_qnt}
\alias{clock_chart_qnt}
\title{Clock Chart with Modified Length}
\usage{
clock_chart_qnt(data, time, crit, high = "red", low = "green")
}
\arguments{
\item{data}{A data frame}

\item{time}{Time in 24 hours HH:MM:SS format in the data set  (19:30:01, for example)}

\item{crit}{The quantitative vector by which hands will be modified and colored.}

\item{high}{The color name for the high values. The default is `red`}

\item{low}{The color name for the high values. The default is `green`.
The color names can be vice versa or other colors, depending on the context.}
}
\value{
A ggplot data clock with modified hands.
}
\description{
This function will plot time of events on a 24 hour clock to show which
events took place at what times. The length and color of the hands are
modified according to a quantitative vector.
}
\examples{
df <- tibble::tibble(hr = sample(0:23, 30, replace = TRUE),
                  mnt = sample(0:59, 30, replace = TRUE),
                  sec = sample(0:59, 30, replace = TRUE),
                  time = paste(hr, mnt, sec, sep = ":"),
                  value = sample(40,30))
clock_chart_qnt(df, time, crit = value)
}
