% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock_chart_qnt.R
\name{clock_chart_qnt}
\alias{clock_chart_qnt}
\title{Clock Chart, Length and Color Modified by a Numeric Variable}
\usage{
clock_chart_qnt(data, time, len, Col, high = "red", low = "green")
}
\arguments{
\item{data}{A data frame}

\item{time}{Time in 24 hours HH:MM:SS format in the data set  (19:30:01, for example)}

\item{len}{The numeric vector by which hands will be modified and colored.}

\item{Col}{The color of line segments and points.}

\item{high}{The color name for the high values. The default is \code{red}}

\item{low}{The color name for the high values. The default is \code{green}.
The color names can be vice versa or other colors, depending on the context.
To use a single color for all lines, use same value for \code{high} and \code{low}}
}
\description{
This function will plot time of events on a 24 hour clock to show which
events took place at what times. The length and color of the hands are
modified according to a numeric vector.
}
\details{
Change the title, subtitle or the caption of the plot with
\code{ggplot2::labs()} .Change the legend title by adding
\code{ggplot2::labs(color = "TITLE")}. Add
or modify legend by \code{theme(legend.position = "POSITION")}; the valid
postion names in \code{ggplot2} are \code{top}, \code{bottom}, \code{right}, and \code{left},
excluding more complex options.
}
\examples{
clock_chart_qnt(data = bdquake, time = hms, len = depth,
Col = mag, high = "red", low = "green")
df <- tibble::tibble(hr = sample(0:23, 50, replace = TRUE),
                  mnt = sample(0:59, 50, replace = TRUE),
                  sec = sample(0:59, 50, replace = TRUE),
                  time = paste(hr, mnt, sec, sep = ":"),
                  value = sample(60,50))
clock_chart_qnt(df, time, len = value, Col = value)
p1 <- clock_chart_qnt(df, time, len = value, Col = value, high = "red", low = "blue")
p1 + ggplot2::theme(legend.position = "right")+
ggplot2::labs(title = "Clock chart of random values")+
ggplot2::labs(color = "Indicator") #Legend Title
}
\seealso{
\code{\link[=clock_chart_col]{clock_chart_col()}} for coloring by a numeric variable,
\code{\link[=clock_chart_qlt]{clock_chart_qlt()}} for coloring by a qualitative variable,
\code{\link[=clock_chart_len]{clock_chart_len()}} for modifying length by a numeric variable,
\code{\link[=clock_chart]{clock_chart()}} for the simplest clock chart
}
