% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{clockplot-package}
\alias{clockplot}
\alias{clockplot-package}
\title{clockplot: Plot Event Times on a 24-Hour Clock}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a novel visualization technique for plotting timestamped events on a 24-hour circular clock face. This is particularly useful for analyzing daily patterns, event clustering, and gaps in temporal data. The package also generalizes this approach to create cyclic charts for other periods, including weekly and monthly cycles, enabling effective event planning and pattern analysis across multiple time frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mahmudstat/clockplot/}
  \item Report bugs at \url{https://github.com/mahmudstat/clockplot/issues}
}

}
\author{
\strong{Maintainer}: Abdullah Al Mahmud \email{almahmud.sbi@gmail.com} (\href{https://orcid.org/0000-0003-2814-8798}{ORCID}) [copyright holder]

}
\keyword{internal}
