% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{clockplot-package}
\alias{clockplot}
\alias{clockplot-package}
\title{clockplot: Plot Event Times on a 24-Hour Clock}
\description{
The primary goal is to plot event times on a 24-hour clock. Additionally it helps you make cyclic charts such a day chart, a week chart, a month chart, or plan events in those periods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mahmudstat/clockplot/}
  \item Report bugs at \url{https://github.com/mahmudstat/clockplot/issues}
}

}
\author{
\strong{Maintainer}: Abdullah Al Mahmud \email{almahmud.sbi@gmail.com} (\href{https://orcid.org/0000-0003-2814-8798}{ORCID}) [copyright holder]

}
\keyword{internal}
