% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{clockplot-package}
\alias{clockplot}
\alias{clockplot-package}
\title{clockplot: Show event time on a 24 hours clock}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
