% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{clockplot-package}
\alias{clockplot}
\alias{clockplot-package}
\title{clockplot: Show Event Time on a 24 Hours Clock and Other Cyclic Plots}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The primary goal of clockplot is to plot event time on a clock chart. Additionally it helps you make a day chart, week chart, or month chart, or plan events in those periods.
}
\keyword{internal}
