% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{conv_hms}
\alias{conv_hms}
\title{Remove second part from time value}
\usage{
conv_hms(df, time)
}
\arguments{
\item{df}{A data set}

\item{time}{a variable in the format \code{HH:MM:SS}}
}
\value{
Time in HH:MM format
}
\description{
This function removes second part from time value
}
\examples{
df <- data.frame(time = c("06:00", "08:00", "17:30"))
conv_hms(df, time)
}
