% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclic_chart.R
\name{cyclic_chart}
\alias{cyclic_chart}
\title{Clyclic Chart}
\usage{
cyclic_chart(df, Period, Value, crit, ColV)
}
\arguments{
\item{df}{A data frame}

\item{Period}{A list of periodical values such as hours of the day, days of
the week, months of the year and so on.}

\item{Value}{A numeric vector with the values corresponding to the \code{Period}}

\item{crit}{A factor variable based on which the bars would be colored.}

\item{ColV}{A character vector with the list of colors for the bars.
You can use this online tool (\url{https://r-charts.com/color-palette-generator/})
to create a beautiful color palette.}
}
\value{
A circular chart showing values in each month of the year.
}
\description{
This function plots values corresponding to random periods such as hours,
days, months and so on.
}
\examples{
# Using package built-in bdtemp data
Col <- c("#bfff00", "#ffff00", "#ffbf00")
cyclic_chart(bdtemp, Period = Month, Value = Temperature,
crit = City, ColV = Col)
}
