% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_chart.R
\name{day_chart}
\alias{day_chart}
\title{Day Chart}
\usage{
day_chart(hvalue, high = "blue", low = "yellow", width = 0.8)
}
\arguments{
\item{hvalue}{A numeric vector having values at each of 24 hours (starts
from 6 am)}

\item{high}{The color name for the high values. The default is `red`}

\item{low}{The color name for the high values. The default is `green`.
The color names can be vice versa or other colors, depending on the context.}
}
\value{
A circular chart showing values at each hour on a 24-hour clock
}
\description{
This function plots values corresponding to each hour on a rose plot.
}
\examples{
value <- sample(15:30,24, replace = TRUE)
day_chart(hvalue = value, high = "blue", low = "yellow", width = 0.8)
}
