% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_chart.R
\name{month_chart}
\alias{month_chart}
\title{Month Chart}
\usage{
month_chart(
  df,
  Days,
  mvalue,
  lgnm = "Value",
  width = 0.9,
  high = "yellow",
  low = "green"
)
}
\arguments{
\item{df}{A data frame containing days of the month and corresponding values.}

\item{Days}{Days of the month; should start from 1 and be between 28 and 31}

\item{mvalue}{A numeric vector having values in each month of the year (starts
from January, obviously)}

\item{lgnm}{Title of legend. The legend position can also be changed by
adding \code{legend.position = "POSITION"}, where the accepted positions are
\code{top}, \code{bottom}, \code{left}, and \code{right}.}

\item{width}{Width of bars}

\item{high}{The color name for the high values. The default is \code{red}}

\item{low}{The color name for the high values. The default is \code{green}.
#' The color names can be vice versa or other colors, depending on the context.}
}
\value{
A circular chart showing values in each month of the year.
}
\description{
This function plots values corresponding to each month on a rose plot.
}
\examples{
Days <- 1:30
expense <- c(sample(15:20, 15, replace = TRUE),
             sample(10:15, 5, replace = TRUE),
             sample(8:10, 5, replace = TRUE),
             sample(5:8, 3, replace = TRUE),
             sample(4:7, 2, replace = TRUE))
dfm <- tibble::tibble(Days, expense)
month_chart(df = dfm, Days = Days, mvalue = expense)
}
