% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_day.R
\name{plan_day}
\alias{plan_day}
\title{Visually Plan Activities on a Day with Events on Each of 24 Hours}
\usage{
plan_day(dwork, width = 1, brdcol = "grey")
}
\arguments{
\item{dwork}{A character vector of length 24, with activity names for each
hour starting from 6 AM and ending at 5 AM the next day.}

\item{width}{Width of bars.}

\item{brdcol}{Color of bar border. Use \code{NA} for no border.}
}
\value{
A \code{ggplot} object, which can be further modified
with \code{ggplot2} functions and themes.
}
\description{
This function plots works corresponding to each hour on a rose plot.
}
\examples{
set.seed(123)
work <- sample(c("Study", "Adda", "Entertainment", "Games", "Exercise", "Meal"),
  size = 24, replace = TRUE
)
plan_day(dwork = work, brdcol = NA)
}
