% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_day.R
\name{plan_day}
\alias{plan_day}
\title{Day Planner}
\usage{
plan_day(dwork, width = 1, brdcol = "grey")
}
\arguments{
\item{dwork}{A character vector having names of work at each of 24 hours (starts
from 6 am)}

\item{width}{Width of bars}

\item{brdcol}{Color of bar border. To have no (transparent color), use \code{NA}}
}
\value{
A circular chart showing work at each hour on a 24-hour clock
}
\description{
This function plots works corresponding to each hour on a rose plot.
}
\examples{
work <- sample(c("Study", "Adda", "Entertainment", "Games", "Exercise", "Meal"),
size = 24, replace = TRUE)
plan_day(dwork = work, brdcol = NA)
}
