% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smsclock}
\alias{smsclock}
\title{Data of Times of Receiving of SMS}
\format{
A data frame with 40 rows and 5 variables:
\describe{
\item{Date}{Date}
\item{time}{Time in HH:MM:SS format}
\item{sender}{Name of the sms sender}
\item{invoked}{Whether the sms is invoked by the user}
\item{type}{The category of the sms; offer, info, ad, service etc.}
...
}
}
\source{
{Phone Messages}
}
\usage{
smsclock
}
\description{
A dataset containing timings of receiving sms on mobile phone.
Real data for a particular day
}
\keyword{datasets}
