% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_chart.R
\name{week_chart}
\alias{week_chart}
\title{Week Chart}
\usage{
week_chart(wvalue, lgnm = "Value", high = "yellow", low = "green")
}
\arguments{
\item{wvalue}{A numeric vector having values on each day, starting from
Saturday}

\item{lgnm}{Title of legend}

\item{high}{The color name for the high values. The default is `red`}

\item{low}{The color name for the high values. The default is `green`.
The color names can be vice versa or other colors, depending on the context.}
}
\value{
A circular chart showing values at each hour on a 24-hour clock
}
\description{
This function plots values corresponding to each day on a rose plot.
}
\examples{
set.seed(10)
wtemp <- sample(10:40,7)
week_chart(wtemp, high = "yellow")+ggplot2::labs(title = "Random Values by Weekday")
}
