% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/year_chart.R
\name{year_chart}
\alias{year_chart}
\title{Year Chart}
\usage{
year_chart(mvalue, lgnm = "Value", width = 0.9, high = "yellow", low = "green")
}
\arguments{
\item{mvalue}{A numeric vector having values in each month of the year (starts
from January, obviously)}

\item{lgnm}{Title of legend. The legend position can also be changed by
adding `legend.position = "top"`, where the accepted positions are
`top`, `bottom`, `left`, and `right`.}

\item{width}{Width of bars}

\item{high}{The color name for the high values. The default is `red`}

\item{low}{The color name for the high values. The default is `green`.
#' The color names can be vice versa or other colors, depending on the context.}
}
\value{
A circular chart showing values in each month of the year.
}
\description{
This function plots values corresponding to each month on a rose plot.
}
\examples{
syltmp <- c(18.4, 20.8, 24.3, 26.0, 26.8, 27.6, 28.0, 28.2, 27.9, 26.7, 23.3, 19.7)
year_chart(mvalue = syltmp)
}
